Started ConvNets
keras_lenet_mnist.py is a quick and dirty implementation of LeNet. Reduced the number of filters so that it can run within a minute in my 8 year old PC with no GPU. :/

keraslenet.py is a good implementation with 32 filters. Ran for 12 epochs and each epoch took ~500 secs to run. Achieved 99.5% accuracy.